Switch to graphql so we can get the actual base sha
The pulls api base sha isn't the head, it's actually the last common sha between the two branches. This made the behavior inconsistent since sometimes it'd be the head and others it wouldn't be.

To fix this we need to use two api calls (one to pulls and one to branch) or one graphql call. The graphql call seems like less points of failure but I'm sure that'll come back and bite me at some point.